powerpc/44x: allow the kernel to be run from a non-zero physical address
CONFIG_PHYSICAL_START always be used to set the LOAD program header
physical address field in the resulting ELF image. And the kernel
also uses this address to set the load and entry address when creating
an uImage. So we need to set this physical address properly when we
have to load the uImage to a non-zero physical address.

Signed-off-by: Kevin Hao <kexin.hao@windriver.com>