powerpc/configs: Address error message when issuing reboot
The default kernel config was enabling CPU Hotplug, which is a feature
not currently supported by the ACP34xx. For this reason, issuing a reboot
or poweroff command was resulting in a "Error taking CPU1 down: -38" error.

This patch disables CONFIG_SUSPEND in order to disable CPU hotplug and
addresses this error message.

Signed-off-by: David Mercado <david.mercado@windriver.com>