mips: have the mips BSPs disable function tracing instead of ftrace
The problem the mips machines apparently ran into was due to
CONFIG_FUNCTION_TRACER et al - no need to disable all of the tracing
infrastructure (CONFIG_FTRACE) to disable that.

Also, the ftrace-disable feature they were using disabled
CONFIG_DEBUG_KERNEL too, which is just a switch to allow other options
to be enabled but doesn't enable anything on itself, so no need for
that either.

Signed-off-by: Tom Zanussi <tom.zanussi@linux.intel.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>