standard.scc: include features/input/input.scc
Include the input.scc to get the CONFIG_INPUT_EVDEV enabled.

The evdev kernel driver is needed to create /dev/input/event* devices.
These devices are used by Xserver to connect to keyboard & mouse kind
of input devices. Without this change some of the BSPs need
  AutoAddDevices = false
in their xorg.conf, which is considered as an undesired hack around
the issue.

Fixes Bug:
[YOCTO #5279]

Signed-off-by: Nitin A Kamble <nitin.a.kamble@intel.com>