meta: Add wifi features
Create a set of wifi features. Create a common fragment for things like
the MAC, CONFIG, and WIRELESS_EXT configs.

Create a fragment for common drivers. Create vendor/class specific
fragments where there is an obvious grouping or where a particular
driver pulls in features that are not generally useful.

Create a complete feature which includes all drivers, but do not move
existing features (such as iwl*), these can be moved under the
features/wifi directory in linux-yocto-dev and forward.

Signed-off-by: Darren Hart <dvhart@linux.intel.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>