minnow: Remove branch statements from minnow scc files
MinnowBoard now builds from standard/base, there is no
need for the branch command in the scc files which create
or reuse a standard/minnow branch. As the goal is to work
from a single branch for all IA BSPs, this is a first step
toward purging the IA BSP branches from linux-yocto.

Signed-off-by: Darren Hart <dvhart@linux.intel.com>