ARM: axxia: Support MSI on both PCIe controllers
Add support for PCIe MSI on both controllers. On AXM5516, PCIE0 has the ability
to signal MSI interrupts on 16 separate lines to the CPU cores, where as PCIE1
only has a single interrupt line that is used for legacy, status/error and MSI.
This patch adds MSI support on a controller with only one interrupt line.

Signed-off-by: Anders Berg <anders.berg@lsi.com>